Update in progress to character select prototype and game start
Created a Character pool which holds all the character models to be selected from.
Created character models which are serializable data a sprite and string name and able to be stored in character pool for easy creating characters to select.
Created a level loading transition for each scene change.
created a character class for gameplay to start brainstorming data for playerprefs.
created a way to change scenes
created a way to select character
just need to implement a working character selection. then ill see if animating makes sense or if making a character appear.